Update rspec-system tests for rspec-system-puppet 1.5
