+++ Gemfile
@@ -2,7 +2,7 @@ source 'https://rubygems.org'
 
 group :development, :test do
   gem 'puppetlabs_spec_helper', :require => false
-  gem 'rspec-system-puppet', '~>1.1', '>=1.1.0'
+  gem 'rspec-system-puppet', '~>1.2', '>=1.2.0'
 end
 
 if puppetversion = ENV['PUPPET_GEM_VERSION']
