+++ spec/system/basic_spec.rb
@@ -4,10 +4,10 @@ require 'spec_helper_system'
 describe "basic tests:" do
   it 'make sure we have copied the module across' do
     # No point diagnosing any more if the module wasn't copied properly
-    system_run("ls /etc/puppet/modules/firewall") do |r|
-      r[:exit_code].should == 0
-      r[:stdout].should =~ /Modulefile/
-      r[:stderr].should == ''
+    shell 'ls /etc/puppet/modules/firewall' do |r|
+      r.stdout.should =~ /Modulefile/
+      r.stderr.should be_empty
+      r.exit_code.should be_zero
     end
   end
 end
