+++ spec/system/purge_spec.rb
@@ -4,8 +4,8 @@ describe "purge tests:" do
   it 'make sure duplicate existing rules get purged' do
     iptables_flush_all_tables
 
-    system_run('/sbin/iptables -A INPUT -s 1.2.1.2')
-    system_run('/sbin/iptables -A INPUT -s 1.2.1.2')
+    shell('/sbin/iptables -A INPUT -s 1.2.1.2')
+    shell('/sbin/iptables -A INPUT -s 1.2.1.2')
     pp = <<-EOS
 class { 'firewall': }
 resources { 'firewall':
@@ -13,13 +13,13 @@ resources { 'firewall':
 }
     EOS
     puppet_apply(pp) do |r|
-      r[:stderr].should == ''
-      r[:exit_code].should == 2
+      r.stderr.should be_empty
+      r.exit_code.should == 2
     end
 
     system_run('/sbin/iptables-save') do |r|
-      r[:stdout].should_not =~ /1\.2\.1\.2/
-      r[:stderr].should == ''
+      r.stdout.should_not =~ /1\.2\.1\.2/
+      r.stderr.should be_empty
     end
   end
 end
